requirements.txt
pandas
psycopg2-binary
openpyxl
python-dotenv
langchain_core
langchain-community
langchain_experimental
langchain-ollama                # install: curl -fsSL https://ollama.com/install.sh | sh
langchain_nvidia_ai_endpoints   # api key: https://build.nvidia.com/meta/llama-3_1-405b-instruct
langchain_huggingface
crewai
crewai_tools
python-pptx
langgraph